Relocate validation handlers
