+++ src/schemaType/ISOTimeType.ts
@@ -1,7 +1,7 @@
 import moment from 'moment';
 import { ISOTimeFormat } from '../constants';
 import { TransformDataError } from '../errors';
-import { handleTypeValidation } from '../utils';
+import { handleTypeValidation } from '../validators';
 import BaseDateType from './BaseDateType';
 import type { ScalarTypeConstructorOptions } from './BaseScalarType';
 import type { SchemaTypeDefinitionBase } from './BaseSchemaType';
