--- src/utils/handleTypeValidation.ts
@@ -1,12 +1,0 @@
-import type { ValidationFunction, Validator } from '../types';
-
-/**
- * Create Validator for type validation
- */
-const handleTypeValidation = (defaultValidator: ValidationFunction): Validator => {
-	const message = 'Property cannot be cast into the defined type';
-
-	return { validator: defaultValidator, message };
-};
-
-export default handleTypeValidation;
