+++ src/utils/index.ts
@@ -1,5 +1,3 @@
 export { default as dummyLogger } from './dummyLogger';
 export { default as ensureArray } from './ensureArray';
 export { default as getFromMvArray } from './getFromMvArray';
-export { default as handleRequiredValidation } from './handleRequiredValidation';
-export { default as handleTypeValidation } from './handleTypeValidation';
