+++ src/validators/handleTypeValidation.ts
@@ -1,0 +1,10 @@
+import type { ValidationFunction, Validator } from '../types';
+
+/** Create Validator for type validation */
+const handleTypeValidation = (validationFn: ValidationFunction): Validator => {
+	const message = 'Property cannot be cast into the defined type';
+
+	return { validator: validationFn, message };
+};
+
+export default handleTypeValidation;
