+++ src/validators/index.ts
@@ -1,0 +1,2 @@
+export { default as handleRequiredValidation } from './handleRequiredValidation';
+export { default as handleTypeValidation } from './handleTypeValidation';
